chore: Add new C++ header libs to geoarrow.hpp and tighten compiler warnings (#118)
The main change here is to lean in to int64 in array_util.hpp, because
it makes it easier to avoid warnings with a variety of uses (plus
interop with Arrow, which is all int64).

+++ src/geoarrow/hpp/array_writer.hpp
@@ -69,9 +69,9 @@ class ArrayBuilder {
 
   template <typename T>
   void AppendToOffsetBuffer(int64_t i, const T& obj) {
-    GEOARROW_THROW_NOT_OK(nullptr,
-                          GeoArrowBuilderOffsetAppend(&builder_, i, obj.data(),
-                                                      static_cast<int64_t>(obj.size())));
+    GEOARROW_THROW_NOT_OK(nullptr, GeoArrowBuilderOffsetAppend(
+                                       &builder_, static_cast<int32_t>(i), obj.data(),
+                                       static_cast<int64_t>(obj.size())));
   }
 
   void AppendCoords(const GeoArrowCoordView* coords, enum GeoArrowDimensions dimensions,
